+++ apps/expo-go/android/vendored/unversioned/react-native-gesture-handler/android/src/main/java/com/swmansion/gesturehandler/react/eventbuilders/ManualGestureHandlerEventDataBuilder.kt
@@ -1,5 +1,0 @@
-package com.swmansion.gesturehandler.react.eventbuilders
-
-import com.swmansion.gesturehandler.core.ManualGestureHandler
-
-class ManualGestureHandlerEventDataBuilder(handler: ManualGestureHandler) : GestureHandlerEventDataBuilder<ManualGestureHandler>(handler)