+++ apps/expo-go/ios/vendored/unversioned/react-native-gesture-handler/ios/RNGHTouchEventType.h
@@ -1,9 +1,0 @@
-#import <Foundation/Foundation.h>
-
-typedef NS_ENUM(NSInteger, RNGHTouchEventType) {
-  RNGHTouchEventTypeUndetermined = 0,
-  RNGHTouchEventTypePointerDown,
-  RNGHTouchEventTypePointerMove,
-  RNGHTouchEventTypePointerUp,
-  RNGHTouchEventTypeCancelled,
-};