+++ docs/ui/components/Footer/NewsletterSignUp.tsx
@@ -42,7 +42,7 @@ export const NewsletterSignUp = () => {
     <div className="flex-1 max-w-[350px] max-md-gutters:max-w-full">
       <CALLOUT theme="secondary" weight="medium" className="flex gap-2 items-center">
         <Mail01Icon className="text-icon-tertiary shrink-0" />
-        Sign up for the Expo's Newsletter
+        Sign up for the Expo Newsletter
       </CALLOUT>
       <form
         className="relative"