+++ packages/@expo/cli/src/start/server/metro/__tests__/externals.test.ts
@@ -1,6 +1,4 @@
-import { vol } from 'memfs';
-
-import { NODE_STDLIB_MODULES, isNodeExternal, setupNodeExternals } from '../externals';
+import { NODE_STDLIB_MODULES, isNodeExternal } from '../externals';
 
 describe('NODE_STDLIB_MODULES', () => {
   it(`works`, () => {
@@ -18,22 +16,5 @@ describe(isNodeExternal, () => {
   it('should return null for non-node modules', () => {
     expect(isNodeExternal('expo')).toBe(null);
     expect(isNodeExternal('expo:fs')).toBe(null);
-  });
-});
-
-describe(setupNodeExternals, () => {
-  afterEach(() => vol.reset());
-  it('should create the correct files', async () => {
-    const projectRoot = '/';
-
-    vol.fromJSON({}, '/');
-
-    await setupNodeExternals(projectRoot);
-
-    expect(Object.keys(vol.toJSON()).length).toBeGreaterThan(42);
-    expect(vol.toJSON()['/.expo/metro/polyfill.native.js']).toBeDefined();
-    expect(vol.toJSON()['/.expo/metro/polyfill.js']).toBeDefined();
-    expect(vol.toJSON()['/.expo/metro/externals/fs/promises/index.js']).toBeDefined();
-    expect(vol.toJSON()['/.expo/metro/externals/assert/index.js']).toBeDefined();
   });
 });